+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is inspired from [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and the versioning aim to respect [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.15.0]
+### Changed
+- updated frictionless (from 4.40.11 to 5.13.1)
+
+### Fixed
+- CSV validation due to header rows and delimiter
+- fixed frictionless update-related stuff
+
 ## [0.14.0] - 2023-05-11
 ### Added
 - SEDOS energysystem structure graph